refactor: make shared utils/constants
