+++ packages/vite/src/node/__tests__/plugins/dynamicImportVar/parse.spec.ts
@@ -2,7 +2,8 @@ import { resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { describe, expect, it } from 'vitest'
 import { transformDynamicImport } from '../../../plugins/dynamicImportVars'
-import { isWindows, normalizePath } from '../../../utils'
+import { normalizePath } from '../../../utils'
+import { isWindows } from '../../../../shared/utils'
 
 const __dirname = resolve(fileURLToPath(import.meta.url), '..')
