+++ packages/vite/src/node/optimizer/scan.ts
@@ -20,7 +20,6 @@ import {
   SPECIAL_QUERY_RE,
 } from '../constants'
 import {
-  cleanUrl,
   createDebugger,
   dataUrlRE,
   externalRE,
@@ -37,6 +36,7 @@ import {
 import type { PluginContainer } from '../server/pluginContainer'
 import { createPluginContainer } from '../server/pluginContainer'
 import { transformGlobImport } from '../plugins/importMetaGlob'
+import { cleanUrl } from '../../shared/utils'
 
 type ResolveIdOptions = Parameters<PluginContainer['resolveId']>[2]
