+++ packages/vite/src/node/plugins/assetImportMetaUrl.ts
@@ -4,13 +4,9 @@ import { stripLiteral } from 'strip-literal'
 import type { Plugin } from '../plugin'
 import type { ResolvedConfig } from '../config'
 import type { ResolveFn } from '../'
-import {
-  injectQuery,
-  isParentDirectory,
-  slash,
-  transformStableResult,
-} from '../utils'
+import { injectQuery, isParentDirectory, transformStableResult } from '../utils'
 import { CLIENT_ENTRY } from '../constants'
+import { slash } from '../../shared/utils'
 import { fileToUrl } from './asset'
 import { preloadHelperId } from './importAnalysisBuild'
 import type { InternalResolveOptions } from './resolve'
