+++ packages/vite/src/node/plugins/importAnalysis.ts
@@ -27,7 +27,6 @@ import {
   normalizeHmrUrl,
 } from '../server/hmr'
 import {
-  cleanUrl,
   createDebugger,
   fsPathFromUrl,
   generateCodeFrame,
@@ -48,10 +47,8 @@ import {
   stripBomTag,
   timeFrom,
   transformStableResult,
-  unwrapId,
   urlRE,
   withTrailingSlash,
-  wrapId,
 } from '../utils'
 import { getFsUtils } from '../fsUtils'
 import { checkPublicFile } from '../publicDir'
@@ -60,6 +57,7 @@ import type { ResolvedConfig } from '../config'
 import type { Plugin } from '../plugin'
 import { shouldExternalizeForSSR } from '../ssr/ssrExternal'
 import { getDepsOptimizer, optimizedDepNeedsInterop } from '../optimizer'
+import { cleanUrl, unwrapId, wrapId } from '../../shared/utils'
 import { throwOutdatedRequest } from './optimizedDeps'
 import { isCSSRequest, isDirectCSSRequest } from './css'
 import { browserExternalId } from './resolve'
