+++ packages/vite/src/node/plugins/preAlias.ts
@@ -9,7 +9,6 @@ import type { Plugin } from '../plugin'
 import { createIsConfiguredAsSsrExternal } from '../ssr/ssrExternal'
 import {
   bareImportRE,
-  cleanUrl,
   isInNodeModules,
   isOptimizable,
   moduleListContains,
@@ -17,6 +16,7 @@ import {
 } from '../utils'
 import { getFsUtils } from '../fsUtils'
 import { getDepsOptimizer } from '../optimizer'
+import { cleanUrl } from '../../shared/utils'
 import { tryOptimizedResolve } from './resolve'
 
 /**
