+++ packages/vite/src/node/server/middlewares/base.ts
@@ -1,10 +1,6 @@
 import type { Connect } from 'dep-types/connect'
-import {
-  cleanUrl,
-  joinUrlSegments,
-  stripBase,
-  withTrailingSlash,
-} from '../../utils'
+import { joinUrlSegments, stripBase, withTrailingSlash } from '../../utils'
+import { cleanUrl } from '../../../shared/utils'
 
 // this middleware is only active when (base !== '/')
