+++ packages/vite/src/node/server/middlewares/static.ts
@@ -7,7 +7,6 @@ import escapeHtml from 'escape-html'
 import type { ViteDevServer } from '../..'
 import { FS_PREFIX } from '../../constants'
 import {
-  cleanUrl,
   fsPathFromId,
   fsPathFromUrl,
   isFileReadable,
@@ -15,12 +14,11 @@ import {
   isInternalRequest,
   isParentDirectory,
   isSameFileUri,
-  isWindows,
   normalizePath,
   removeLeadingSlash,
-  slash,
   withTrailingSlash,
 } from '../../utils'
+import { cleanUrl, isWindows, slash } from '../../../shared/utils'
 
 const knownJavascriptExtensionRE = /\.[tj]sx?$/
