+++ packages/vite/src/node/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["./", "../runtime", "../dep-types", "../types"],
+  "include": ["./", "../runtime", "../dep-types", "../types", "constants.ts"],
   "exclude": ["../**/__tests__"],
   "compilerOptions": {
     "lib": ["ESNext", "DOM"],
